+++ i3df/src/lib.rs
@@ -269,7 +269,7 @@ pub fn decode_array_texture(mut reader: impl Read) -> Result<Vec<Texture>, Error
     let mut array = Vec::new();
 
     for _ in 0..item_count {
-        let file_id = reader.read_u64::<LittleEndian>()? as u32;
+        let file_id = reader.read_u64::<LittleEndian>()? as f64;
         let width = reader.read_u16::<LittleEndian>()?;
         let height = reader.read_u16::<LittleEndian>()?;
         let _reserved = reader.read_u32::<LittleEndian>()?;
